Split preview panel into its own class
This is much much cleaner. Less state can get muddied between the two
classes, and inputs and outputs are much more obvious.
