+++ Assets/RedBlueGames/MulliganRenamer/Editor/MulliganRenamerPreviewPanel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c809b109d043420daae293a006b5c53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
